# backend/05_openapi_final.yaml
openapi: 3.1.0
info:
  title: Kaveri Stays API
  version: "1.0.0"
  summary: Reveal specification for the Kaveri Stays API assignment.
  description: |
    # Read this before you read the paths

    This is ONE reasonable design. It is not the only one. Several of the choices
    below are genuinely arguable and the Interlude expects you to argue with them.
    A few are not arguable, and those are marked.

    ## Conventions this specification commits to

    **Dates.** Every date is ISO 8601 `YYYY-MM-DD`. A stay is the half-open interval
    `[check_in, check_out)`. The check-out date is NOT part of the stay. This is what
    makes same-day turnover work (rule 5), and it is NOT arguable.

    **Money.** Every amount is a decimal string, never a JSON number. `"4500.00"`,
    not `4500.0`. Floats lose money.

    **Empty collections.** A list endpoint with no matching rows returns 200 and an
    empty `items` array. It never returns 404. 404 means the container does not exist,
    not that it is empty.

    **401 versus 403.** 401 means the server does not know who you are: no token, an
    expired token, a forged token. 403 means it knows exactly who you are and you may
    not do this. This distinction is NOT arguable.

    **Request bodies.** Every request schema sets `additionalProperties: false`. A
    field you did not declare is rejected rather than ignored. This is what defeats
    attacks 8.2 and 8.8 structurally rather than by remembering to check.

...

...

...

...

...

  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      summary: Rotate a refresh token.
      description: |
        Refresh tokens are single use. Presenting one that has already been rotated
        is 401, and a correct implementation also invalidates the whole token family,
        because reuse means the token leaked. This is attack 8.6.
      security: []
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: New token pair issued; the presented refresh token is now dead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /bookings/{booking_id}/review:
    post:
      tags: [reviews]
      operationId: createReview
      summary: Review a completed stay.
      description: |
        Task 5.8. Rule 9: only after checkout, only once, rating 1 to 5.

        Three distinct failures with three distinct codes. Reviewing someone else's
        stay is 404, consistent with `GET /bookings/{id}`. Reviewing your own stay
        before you have checked out is 403, because you personally are not yet
        permitted to do this. A second review on the same booking is 409, because it
        collides with a row that already exists. A rating of 9 is 422.
      x-roles: [guest]
      parameters:
        - $ref: '#/components/parameters/BookingIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewCreate' }
      responses:
        '201':
          description: Review recorded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: The stay is not checked out yet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: This booking already has a review.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /properties/{property_id}/reviews:
    get:
      tags: [reviews]
      operationId: listPropertyReviews
      summary: Reviews for a property.
      description: Guest names are shortened. Reviews never expose a guest email.
      security: []
      x-roles: [public]
      parameters:
        - $ref: '#/components/parameters/PropertyIdPath'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Reviews.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewPage' }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------- reports ---
  /reports/occupancy:
    get:
      tags: [reports]
      operationId: reportOccupancy
      summary: Occupancy percentage per property per month.
      description: |
        A manager omitting `property_id` gets their own property. A manager naming a
        property that is not theirs gets 403, not an empty result. This is attack 8.7.

        The denominator is room-nights available, counting every room in the property
        whether or not it was ever booked.
      x-roles: [manager, owner]
      parameters:
        - $ref: '#/components/parameters/ReportPropertyId'
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: Occupancy rows.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/OccupancyRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /reports/adr:
    get:
      tags: [reports]
      operationId: reportAdr
      summary: Average daily rate per property per month.
      x-roles: [manager, owner]
      parameters:
        - $ref: '#/components/parameters/ReportPropertyId'
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: ADR rows.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RateMetricRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/revpar:
    get:
      tags: [reports]
      operationId: reportRevpar
      summary: Revenue per available room, per property per month.
      x-roles: [manager, owner]
      parameters:
        - $ref: '#/components/parameters/ReportPropertyId'
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: RevPAR rows.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RateMetricRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/revenue:
    get:
      tags: [reports]
      operationId: reportRevenue
      summary: Revenue per property per month, across all properties.
      description: |
        Owner only. A manager receives 403 even for their own property, because this
        endpoint's shape is cross-property by definition. Managers use the three
        endpoints above.
      x-roles: [owner]
      parameters:
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: Revenue rows, with a grand total.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, grand_total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RevenueRow' }
                  grand_total: { type: string, examples: ["4820500.00"] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # --------------------------------------------------------------- guests ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token in `Authorization: Bearer <token>`. Short lived. Obtain a new one
        from `/auth/refresh` rather than by logging in again.
  parameters:
    PropertyIdPath:
      name: property_id
      in: path
      required: true
      schema: { type: integer }
    BookingIdPath:
      name: booking_id
      in: path
      required: true
      schema: { type: integer }
    LimitParam:
      name: limit
      in: query
      description: Page size. Capped server-side; an unbounded list endpoint is a defect.
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
    ReportPropertyId:
      name: property_id
      in: query
      description: Omit to mean "the property I manage". Owners may name any property.
      schema: { type: integer }
    ReportFrom:
      name: from
      in: query
      required: true
      schema: { type: string, format: date }
    ReportTo:
      name: to
      in: query
      required: true
      schema: { type: string, format: date }
  responses:
    Unauthorized:
      description: No token, an expired token, or a token that does not verify.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: unauthenticated, message: Authentication required. }
    Forbidden:
      description: Authenticated, and not permitted to do this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: forbidden, message: You do not have access to this resource. }
    NotFound:
      description: No such resource, or none you are permitted to know about.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: not_found, message: Not found. }
...
    Unprocessable:
      description: |
        The request parsed and a field is unacceptable. Includes guest counts above
        the room type maximum, check-out on or before check-in, ratings outside 1 to 5,
        and any field not declared in the request schema.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: validation_failed
              message: Request validation failed.
              detail:
                - field: guests
                  reason: Exceeds maximum occupancy of 2 for room type Standard.
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      description: |
        Every non-2xx response in this API uses this shape. One handler produces it.
        If you are constructing this dict in more than one module, consolidate.
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable, snake_case. Clients branch on this.
              examples: [room_unavailable, illegal_transition, validation_failed]
            message:
              type: string
              description: Human-readable. Safe to show a guest. Never a raw Postgres message.
            detail:
              description: Optional structured context. Never leaks table or column names.
            request_id:
              type: string
              description: Correlates this response with your logs.
    PageMeta:
      type: object
      additionalProperties: false
      required: [limit, offset, total]
      properties:
        limit:  { type: integer }
        offset: { type: integer }
        total:  { type: integer, description: Total matching rows, ignoring pagination. }

    # ---- enums ----
...
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    TokenPair:
      type: object
      additionalProperties: false
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:  { type: string }
        refresh_token: { type: string }
        token_type:    { type: string, const: bearer }
        expires_in:    { type: integer, description: Access token lifetime in seconds. }
...
    ReviewCreate:
      type: object
      additionalProperties: false
      required: [rating]
      properties:
        rating:  { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 2000 }
    Review:
      type: object
      additionalProperties: false
      required: [id, booking_id, rating, created_at]
      properties:
        id:          { type: integer }
        booking_id:  { type: integer }
        rating:      { type: integer, minimum: 1, maximum: 5 }
        comment:     { type: string }
        guest_name:  { type: string, description: Shortened. Never the full record, never the email. }
        created_at:  { type: string, format: date-time }
    ReviewPage:
      type: object
      additionalProperties: false
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Review' }
        meta: { $ref: '#/components/schemas/PageMeta' }

    # ---- reports ----
    OccupancyRow:
      type: object
      additionalProperties: false
      required: [property_id, month, room_nights_available, room_nights_sold, occupancy_pct]
      properties:
        property_id:           { type: integer }
        property_name:         { type: string }
        month:                 { type: string, examples: ["2025-12"] }
        room_nights_available: { type: integer }
        room_nights_sold:      { type: integer }
        occupancy_pct:         { type: string, examples: ["78.40"] }
    RateMetricRow:
      type: object
      additionalProperties: false
      required: [property_id, month, value]
      properties:
        property_id:   { type: integer }
        property_name: { type: string }
        month:         { type: string, examples: ["2025-12"] }
        value:         { type: string, examples: ["6420.00"] }
    RevenueRow:
      type: object
      additionalProperties: false
      required: [property_id, month, revenue]
      properties:
        property_id:   { type: integer }
        property_name: { type: string }
        month:         { type: string, examples: ["2025-12"] }
        revenue:       { type: string, examples: ["1284000.00"] }

    # ---- guests ----
...